requirements.txt
aiogram>=3.0
google-genai
python-dotenv
httpx